Criação da classe CovidDataController. Páginas movidas para dentro da pasta pages.
