+++ lib/presentation/pages/search/widget/widget_graphic.dart
