+++ lib/presentation/pages/search/widget/widget_informations.dart
